requirements.txt
argcomplete==1.9.2
base58==0.2.4
biopython==1.70
boto3==1.5.20
botocore==1.8.34
certifi==2018.1.18
cfn-flip==1.0.0
chardet==3.0.4
click==6.7
DateTime==4.2
docutils==0.14
durationpy==0.5
Flask==0.12.2
future==0.16.0
hjson==3.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.19.0
MarkupSafe==1.0
numpy==1.14.0
pandas==0.22.0
placebo==0.8.1
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.3
PyYAML==3.12
regex==2018.1.10
requests==2.18.4
s3transfer==0.1.12
six==1.11.0
titlecase==0.12.0
toml==0.9.4
tqdm==4.19.1
troposphere==2.1.2
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.14.1
wsgi-request-logger==0.4.6
xlrd==1.1.0
zappa==0.45.1
zope.interface==4.4.3
